adding some code

+++ src/components/Profile/DeleteProfileButton.jsx
@@ -2,12 +2,20 @@ import userCalls from "../../../lib/user-api"
 
 const DeleteProfileButton = ({userId}) => {
     const handleDeleteProfileClick = async () => {
-        await userCalls.deleteUser(userId)
+        try {
+            const confirm = window.prompt('Are you sure you want to delete your account? ')
+            if (confirm === 'yes') {
+                await userCalls.deleteUser(userId)
+                alert('account deleted succesfully')
+            }
+        } catch (error) {
+            console.log({error: error.message})
+        }
     }
     
     return (
         <>
-            <button onClick={() => {handleDeleteProfileClick}}>Delete Profile</button>
+            <button onClick={handleDeleteProfileClick}>Delete Profile</button>
         </>
     )
 }
